Rewrite Session Store blocking support using new APIs.
We now filter the JSON data written to the store instead of reimplementing the
component used to create it. This is slightly less fragile than what we used
to do, but perhaps not by much. If the structure of the session store changes
substantially, we'll break. From the comments, it does look like they try to
keep some level of compatibility with previous formats..