Bug #2148: Make the ref spoofing more uniform in behavior
This may not be the final form we want for it to take, see FIXME in the
comments.

Also fix a potential issue with unescaped .'s in regex match strings. I'm
really not sure we can call this ref spoofing protection a real security
feature, though, so it's not like this matters a whole bunch. But might as
well fix it, too.