Bug #3580: Fix hotmail (but only for TBB users)
Hotmail appears to have an optimization strategy that involved pre-loading all
of the scripts involved in their site as object tags. When the user navigates
between pages, the appropriate object tags get converted into script tags and
executed.

The problem for us is that docSell.allowPlugins is implemented as a content
policy that blocks object tags for a given page. We use this API because the toggle
model requires the ability to do per-page plugin control.

Firefox 3.6 introduced a global plugin control API that allows us to
enable/disable indivual plugins, but this was unfit for use in the toggle
situation where tabs might sit around in the background.

However, we can use it for TBB. This patch switches us to using the plugin
manager API, in TBB only.