Bug #3337: Fetch check.tp.o update page if needed
Uses https://check.torproject.org/RecommendedTBBVersions as the version list.
If this url is 404, the feature is disabled. Otherwise, we check to see if
the current value of torbrowser.version is present in the returned JSON list.
If so, we fetch the normal check homepage. Otherwise, we fetch the upgrade
notice.