Bug 4197: Torbutton formfill blocking can't be undone
We still require a toggle, which is very difficult for TBB users. Filed #4198
for that, as several prefs are involved, and we probably want to wait for
bug #3100 anyways.

Also, we should set signon.autoformFill to false during Tor usage regardless
of this pref, for unlinkability. So we now do so.